@@ -12,11 +12,10 @@
         "symfony/asset": "^2.8|^3.0|^4.0",
         "symfony/browser-kit": "^2.8|^3.0|^4.0",
         "symfony/css-selector": "^2.3|^3.0|^4.0",
-        "symfony/swiftmailer-bundle": "^2.3.3|^3.0",
+        "symfony/swiftmailer-bundle": "^2.5|^3.0",
         "symfony/templating": "^2.8|^3.0|^4.0",
         "symfony/twig-bundle": "^2.8|^3.0|^4.0",
-        "twilio/sdk": "^5.0",
-        "php-mock/php-mock-prophecy": "^0.0.2"
+        "twilio/sdk": "^5.0"
     },
     "conflict": {
         "symfony/swiftmailer-bundle": "<2.3.3",
@@ -34,7 +33,8 @@
         }
     },
     "config": {
-        "bin-dir": "bin/"
+        "bin-dir": "bin/",
+        "sort-packages": true
     },
     "extra": {
         "branch-alias": {
