@@ -1,5 +1,5 @@
-<?php
+<?php declare(strict_types=1);
 
 require 'vendor/autoload.php';
 
-class_alias(\PHPUnit\Framework\TestCase::class, \PHPUnit_Framework_TestCase::class);
+\class_alias(\PHPUnit\Framework\TestCase::class, \PHPUnit_Framework_TestCase::class);
